Optimize traversals through diagnostics.
Code is a little bit uglier but we have gained a lot of performance
for certain use-cases (e.g. when there are a lot of diagnostics).

Some preliminary tests show that running the integration test suite
on 'cppcheck' take around 260 seconds to finish. With this optimization
new figure is around 180 seconds.