We need a new table to store diagnostics details.
Up until now those details (diagnostics children) were stored in the
same table as their parent diagnostics. However, that strategy falls short
whenever we want to sort the (parent) diagnostics by some criteria. As soon
as we sort the records by some criteria, we are basically 'unlinking' the
diagnostics children from their respective parents and it is not possible
to reconstruct the parent anymore.

That is why we need a new table where we will store those additional
details so we can extract them whenever is needed.