Re-implement the code-complete logic.
In order to trigger code-complete on modified TUnit, we do _NOT_ have
to (re-)parse the TUnit first as it was believed. This was needlessly
decreasing the performance of code-complete functionality because
libclang implemention of code-complete already handles this use-case
for us via unsaved-files property. Unfortunatelly, this is not very clear
from official documentation.

Default parsing options are now also expanded to a bigger list of flags.
This is to avoid situations where code-complete service fetches TUnit from
cache (which is basically all of the time?), but because that particular
TUnit is parsed (and inserted into the cache) by some other service, it is
highly probable that it will be missing the parsing flags which actually
make code-complete feature more performant (e.g. PARSE_CACHE_COMPLETION_RESULTS,
PARSE_PRECOMPILED_PREAMBLE, ...). So, let's unify those flags to the most
common denominator.