Use appropriate compiler args for modified TUnits.
Compiler args for modified TUnits have to be tweaked in order to get it
parsed right. This was previously (and implicitly) handled by
'CompilerArgs::get()' but error has been introduced once we started caching
the compiler args into the TUnit cache. We were caching only the original
compiler args which resulted in error once modified TUnits started reading
those ones (without the tweak) in order to trigger the parse operation.

Now, we decouple the interface in 'CompilerArgs' so we can massage the compiler
args whenever we need to. It's a bit more responsibility on the client code
but it gives us great possibility of not being required to read compiler args
from JSON database all over and over again (which is expensive if it is done
basically on any event).