Instantiate server with a target. Pass project_root_directory to CxxdConfigParser.
We need a valid target to run the server. If target does not exist or
it's not valid, the best thing we can do is to bail out. In particular,
this means that user did not provide compilation-database or compile-flags
configuration.

We could theoretically continue (as we did before) but with unsatisfying
results.