Implement API for fetching and handling target-specific configurations.
Configurations are to be provided as:
 1. compilation-databases
 2. simple txt files containing build flags, include paths, etc.

There are 2 ways we enable users to provide configurations:
 1. Explicit
 2. Auto-discovery

In explicit, we allow users to provide configurations under which they can
define as many 'targets' as they wish. Each target is mapped to a single
configuration file, be it a compilation-database or compile-flags file.
This is crucial in order to get the best (parsing) results depending on the
target/configuration you want to use for given project. E.g.
  Debug, Release, RelWithDbg, <any_arbitrary_config>, etc.

In auto-discovery, we allow users to provide search-paths which will be
used in order to detect if any configuration exists. This is to simplify
the usage for simple projects. So, not necessarily recommended.